requirements.txt
# Microsoft Agent Framework
agent-framework-a2a==1.0.0b260128
agent-framework-ag-ui==1.0.0b260128
agent-framework-azure-ai==1.0.0b260128
agent-framework-core==1.0.0b260128
agent-framework-declarative==1.0.0b260128
agent-framework-devui==1.0.0b260128
agent-framework-mem0==1.0.0b260128
azure-ai-evaluation==1.14.0
# Requirements File
a2a-sdk==0.3.22
ag-ui-protocol==0.1.10
fastapi==0.128.0
fastmcp==2.14.4
httpx==0.28.1
openai==2.16.0
pydantic==2.12.5
python-dotenv==1.2.1
pytz==2025.2
pandas==2.3.3
uvicorn==0.40.0
uv